requirements.txt
langgraph
langchain_openai
langchain_google_genai
langchain_core
langchain_community
python-dotenv
langgraph-checkpoint-postgres
streamlit
psycopg2
faiss-cpu
sentence-transformers
rapidfuzz